feat: implement service layer with repository integration
- Completed PostServiceIMPL with CRUD, update, search and category filtering
- Completed CategoryServiceIMPL with pagination and search functionality
- Completed CommentServiceIMPL with status management and post-based queries
- Completed TagServiceIMPL with name search and post relationship methods
- Added proper dependency injection through constructors
- Added business logic for comment status defaulting
- Integrated all repository methods with proper error handling

+++ src/main/java/org/example/tblog/service/CommentService.java
@@ -4,7 +4,7 @@ import org.example.tblog.model.Comment;
 import org.example.tblog.model.CommentsStatus;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import java.util.List;
+
 import java.util.Optional;
 
 public interface CommentService {
@@ -13,7 +13,7 @@ public interface CommentService {
     Optional<Comment> findById(Integer id);
     void deleteById(Integer id);
     Page<Comment> findByPostId(Integer postId, Pageable pageable);
-    Long countByPostId(Integer postId);
+    int countByPostId(Integer postId);
     Page<Comment> findByStatus(CommentsStatus status , Pageable pageable);
     Page<Comment> findByPostIdAndStatus(Integer postId, CommentsStatus status , Pageable pageable);
     Comment updateStatus(Integer commentId, CommentsStatus newStatus);
